Improve secret share; add ascii::write

+++ include/data/crypto/secret_share.hpp
@@ -12,13 +12,13 @@
 namespace data::crypto {
     
     struct secret_share {
-        uint32 Index;
+        byte Index;
         bytes Data;
     };
     
-    cross<secret_share> secret_share_split(CryptoPP::RandomNumberGenerator&, const bytes& message, uint32 max, uint32 threshold);
+    cross<secret_share> secret_share_split(CryptoPP::RandomNumberGenerator&, const bytes& message, byte max, byte threshold);
     
-    bytes secret_share_merge(cross<secret_share>, uint32 threshold);
+    bytes secret_share_merge(cross<secret_share>, byte threshold);
     
 }
